rag/requirements.txt
chromadb>=0.4.22
sentence-transformers>=2.2.2
tiktoken>=0.5.2
pyyaml>=6.0.1
markdown>=3.5.1
python-frontmatter>=1.0.0
